requirements.txt
# Inkle.ai Multi-Agent Tourism System - 2025 Compatible Versions
streamlit==1.36.0
langchain==0.1.20
langchain-openai==0.1.10
langchain-community==0.0.30
langgraph==0.0.35
openai==1.3.8
requests==2.31.0
python-dotenv==1.0.0
geopy==2.4.1
pydantic==2.5.0
fastapi==0.104.1
